+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ All notable changes to the ``topology`` project will be documented in this file.
 
 Added
 =====
+- Added endpoint ``POST v3/interfaces/{interface_id}/tag_ranges`` to set ``tag_ranges`` to interfaces.
+- Added endpoint ``DELETE v3/interfaces/{interface_id}/tag_ranges`` to delete ``tag_ranges`` from interfaces.
 
 Deprecated
 ==========